requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.5.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
numpy==1.22.1
oauthlib==3.1.1
pandas==1.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.8
